#2 renamed constants file with proper spelling
