+++ src/libs/editor-theme.ts
@@ -1,4 +1,4 @@
-import { BOLD_VALUE, ITALIC_VALUE } from './contants';
+import { BOLD_VALUE, ITALIC_VALUE } from './constants';
 import {
   UNDERLINE_STYLE,
   STRIKETHROUGH_STYLE,
@@ -21,6 +21,24 @@ const theme = {
     underline: UNDERLINE_STYLE,
     strikethrough: STRIKETHROUGH_STYLE,
     underlineStrikethrough: UNDERLINE_STRIKETHROUGH_STYLE,
+  },
+
+  list: {
+    ul: 'list-disc pl-6 space-y-1',
+    ol: 'list-decimal pl-6 space-y-1',
+    ulDepth: ['list-disc', 'list-[circle]', 'list-[square]', 'list-disc'],
+    olDepth: ['list-decimal', 'list-[lower-alpha]', 'list-[lower-roman]', 'list-decimal'],
+
+    listitem: 'my-1 marker:text-base-content marker:opacity-70',
+
+    nested: {
+      list: 'pl-6',
+      listitem: 'my-1',
+    },
+
+    checklist: 'list-none pl-2',
+    listitemChecked: 'line-through text-base-content/60',
+    listitemUnchecked: 'text-base-content',
   },
 };
